chore: build and release
