+++ .github/workflows/build.yml
@@ -1,0 +1,42 @@
+---
+name: build
+
+on:
+  push:
+    branches:
+      - main
+    paths-ignore:
+      - "**.md"
+      - ".gitignore"
+      - "resources/**"
+  pull_request:
+    branches:
+      - main
+    paths-ignore:
+      - "**.md"
+      - ".gitignore"
+      - "resources/**"
+
+concurrency:
+  group: build-${{ github.event.pull_request.number || github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  build:
+    runs-on: ubuntu-24.04
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4.2.2
+      - name: Setup Go
+        uses: actions/setup-go@v5.3.0
+        with:
+          go-version-file: go.mod
+      - name: Ensure go.mod is already tidied
+        run: go mod tidy && git diff --no-patch --exit-code
+      - name: Build with Goreleaser
+        uses: goreleaser/goreleaser-action@v6.2.1
+        with:
+          version: v2.7.0
+          args: release --snapshot --skip=publish --clean
+        env:
+          PRIVATE_ACCESS_TOKEN: placeholder
