+++ Dockerfile
@@ -1,15 +1,6 @@
-FROM golang:1.16
-MAINTAINER Frank R <12985912+fritchie@users.noreply.github.com>
+FROM ubuntu:24.04
 
-RUN apt-get update
-RUN apt-get -y install nvme-cli
-
-WORKDIR /go/src/nvme_exporter
-COPY . .
-
-RUN go get -d -v ./...
-RUN go install -v ./...
+COPY nvme_exporter /usr/bin/nvme_exporter
 
 EXPOSE 9998
-
-CMD [ "nvme_exporter" ]
+ENTRYPOINT ["/usr/bin/nvme_exporter"]
