+++ resources/scripts/post-remove.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+set -e
+
+userdel -f nvme_exporter || true
+
+systemctl daemon-reload
