+++ resources/scripts/pre-remove.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+set -e
+
+systemctl stop nvme_exporter.service || true
+systemctl disable nvme_exporter.service || true
+
+systemctl daemon-reload
